+++ Frontend_WhatToWeEat/src/App.vue
@@ -1,19 +1,12 @@
 <script setup>
 // This starter template is using Vue 3 <script setup> SFCs
 // Check out https://vuejs.org/api/sfc-script-setup.html#script-setup
-import LoginPage from './components/LoginPage.vue'
-import RegisterPage from './components/RegisterPage.vue';
-
-
-
 
 </script>
 
 <template>
-  <div class="h-screen bg-red-50">
-    <LoginPage/>
-  </div>
-
+  <h1> I am the HomePage</h1>
+  <router-view></router-view>
   
 
 </template>
