+++ Frontend_WhatToWeEat/src/main.js
@@ -1,13 +1,11 @@
 import { createApp } from 'vue'
+import router from './router'; 5466
 import App from './App.vue'
 import './index.css'
 
 
-
 const app = createApp(App)
 
-// const meta = document.createElement('meta')
-// meta.name = 'naive-ui-style'
-// document.head.appendChild(meta)
+app.use(router)
 
 app.mount('#app')
